Refactor feature. Add hooks for maintainability
