+++ src/components/collect-pokemon.tsx
@@ -6,46 +6,25 @@ import { pokemonABI } from '@/abis/CollectPokemon';
 import { useWriteContract, useAccount } from 'wagmi';
 import { wagmiConfig } from '@/config';
 import { toast } from './ui/use-toast';
+import useWalletConnect from '@/hooks/useWalletConnect';
+import useCollectPokemon from '@/hooks/useCollectPokemon';
 
 type CollectButtonProps = {
   pokemonId: number;
   className?: string;
   isCollected: { result: string; status: string };
 };
+
 export function CollectButton({ pokemonId, className, isCollected }: CollectButtonProps) {
-  const { address: ownerAddress } = useAccount();
-  const { data: hash, writeContract, isError, isSuccess, error, isPending } = useWriteContract({ config: wagmiConfig });
+  const { handleCollect, isPending, collectedInSession } = useCollectPokemon(pokemonId);
 
-  const [collectedInSession, setCollectedInSession] = useState(false);
+  let buttonText = 'Collect';
+  if (isCollected?.result !== undefined || collectedInSession) {
+    buttonText = 'Collected';
+  } else if (isPending) {
+    buttonText = 'Collecting';
+  }
 
-  useEffect(() => {
-    if (isError && error) {
-      toast({ title: 'Error', description: error.message, variant: 'destructive' });
-    } else if (isSuccess && hash) {
-      toast({ title: 'Success', description: `Transaction Hash: ${hash}` });
-      setCollectedInSession(true);
-    }
-  }, [isError, isSuccess, error, hash, pokemonId]);
-
-  const handleCollect = async () => {
-    try {
-      writeContract({
-        address: process.env.NEXT_PUBLIC_POKEMON_CONTRACT! as `0x${string}`,
-        abi: pokemonABI,
-        functionName: 'collectPokemon',
-        args: [ownerAddress, BigInt(pokemonId)],
-      });
-      const collectedPokemons = JSON.parse(localStorage.getItem('collectedPokemons') || '[]');
-      if (!collectedPokemons.includes(pokemonId)) {
-        localStorage.setItem('collectedPokemons', JSON.stringify([...collectedPokemons, pokemonId]));
-        setCollectedInSession(true);
-      }
-    } catch (error) {
-      console.error('Collection Error:', error);
-    }
-  };
-
-  const buttonText = isCollected?.result !== undefined || collectedInSession ? 'Collected' : 'Collect';
   return (
     <>
       <Button variant="obolPrimary" className={className} onClick={handleCollect} disabled={isCollected?.result !== undefined || isPending || collectedInSession}>
