+++ src/hooks/useWalletConnect.ts
@@ -1,0 +1,13 @@
+import { useAccount, useConnect } from 'wagmi';
+export default function useWalletConnect() {
+  const { connect, connectors } = useConnect();
+  const { isConnected } = useAccount();
+
+  const connectWallet = async () => {
+    if (!isConnected) {
+      await connect({ connector: connectors[0] });
+    }
+  };
+
+  return { connectWallet, isConnected };
+}
